requirements.txt
pandas
psycopg2-binary
apache-airflow
pytest
boto3
mock
pyarrow
polars